Refactor PieWidgetUI
