+++ packages/react-ui/src/widgets/PieWidgetUI.js
@@ -2,13 +2,8 @@ import React, { useMemo, useRef, useEffect, useState, useCallback } from 'react'
 import PropTypes from 'prop-types';
 import ReactEcharts from '../custom-components/echarts-for-react';
 import { useTheme } from '@material-ui/core';
-import {
-  getChartSerie,
-  areChartPropsEqual,
-  disableSerie,
-  setColor,
-  defaultTooltipFormatter
-} from './utils/chartUtils';
+import { disableSerie, setColor } from './utils/chartUtils';
+import { processFormatterRes } from './utils/formatterUtils';
 
 function PieWidgetUI({
   name,
@@ -25,27 +20,83 @@ function PieWidgetUI({
 }) {
   const theme = useTheme();
   const [showLabel, setShowLabel] = useState(true);
+  const [showTooltip, setShowTooltip] = useState(true);
+  const colorByCategory = useRef({});
 
+  // Reset colorByCategory when colors changes
+  useEffect(() => {
+    colorByCategory.current = {};
+    // Spread colors array to avoid reference problems
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [...colors]);
+
+  const dataWithColor = useMemo(() => {
+    return data.map(processDataItem(colorByCategory, colors, theme));
+  }, [data, colors, theme]);
+
+  // Tooltip
   const tooltipOptions = useMemo(
     () => ({
-      trigger: 'item',
+      show: showTooltip,
       showDelay: 1000,
       transitionDuration: 0,
       backgroundColor: theme.palette.other.tooltip,
-      textStyle: {
-        color: theme.palette.common.white
-      },
+      textStyle: { color: theme.palette.common.white },
       confine: true,
-      ...(tooltipFormatter
-        ? {
-            formatter: (params) =>
-              tooltipFormatter([{ ...params, valueFormatter: formatter }])
-          }
-        : {})
+      formatter:
+        !!tooltipFormatter && ((params) => tooltipFormatter({ ...params, formatter }))
     }),
-    [formatter, theme.palette.common.white, theme.palette.other.tooltip, tooltipFormatter]
+    [
+      showTooltip,
+      formatter,
+      theme.palette.common.white,
+      theme.palette.other.tooltip,
+      tooltipFormatter
+    ]
   );
 
+  // Legend
+  const legendOptions = useMemo(
+    () => ({
+      selectedMode: false,
+      type: 'scroll',
+      left: theme.spacing(1),
+      bottom: -theme.spacing(0.5),
+      itemGap: theme.spacing(3),
+      icon: 'circle',
+      itemWidth: theme.spacing(1),
+      itemHeight: theme.spacing(1),
+      // TODO: as prop?
+      formatter: (name) => name.toUpperCase(),
+      textStyle: {
+        ...theme.typography.charts,
+        color: theme.palette.text.primary,
+        lineHeight: 1,
+        verticalAlign: 'bottom',
+        padding: [0, 0, 0, theme.spacing(0.5)]
+      },
+      inactiveColor: theme.palette.text.disabled,
+      pageIcons: {
+        horizontal: [
+          'path://M15.41 7.41 14 6 8 12 14 18 15.41 16.59 10.83 12z',
+          'path://M9 16.59 13.3265857 12 9 7.41 10.3319838 6 16 12 10.3319838 18z'
+        ]
+      },
+      pageIconSize: theme.spacing(1.5),
+      pageIconColor: theme.palette.text.secondary,
+      pageIconInactiveColor: theme.palette.text.disabled,
+      pageTextStyle: {
+        fontFamily: theme.typography.charts.fontFamily,
+        fontSize: theme.spacing(1.5),
+        lineHeight: theme.spacing(1.75),
+        fontWeight: 'normal',
+        color: theme.palette.text.primary
+      }
+    }),
+    [theme]
+  );
+
+  // Series
   const labelOptions = useMemo(
     () => ({
       formatter: '{per|{d}%}\n{b|{b}}',
@@ -70,22 +121,22 @@ function PieWidgetUI({
     [theme]
   );
 
-  // Series
   const seriesOptions = useMemo(
     () => [
       {
         type: 'pie',
+        name,
         animation,
-        data: data.map((item) => {
+        data: dataWithColor.map((item) => {
           // Avoid modifying data item
-          const clonedItem = {
-            ...item,
-            name: labels?.[item.name] || item.name,
-            key: item.name
-          };
+          const clonedItem = { ...item };
+
+          if (labels?.[clonedItem.name]) {
+            clonedItem.name = labels[clonedItem.name];
+          }
 
           const disabled =
-            selectedCategories?.length && !selectedCategories.includes(clonedItem.key);
+            selectedCategories?.length && !selectedCategories.includes(clonedItem.name);
 
           if (disabled) {
             disableSerie(clonedItem, theme);
@@ -97,6 +148,8 @@ function PieWidgetUI({
           return clonedItem;
         }),
         radius: ['74%', '90%'],
+        selectedOffset: 0,
+        hoverOffset: 5,
         bottom: theme.spacing(2.5),
         label: { show: showLabel, ...labelOptions },
         emphasis: {
@@ -104,7 +157,16 @@ function PieWidgetUI({
         }
       }
     ],
-    [animation, data, labels, selectedCategories, theme, showLabel, labelOptions]
+    [
+      name,
+      animation,
+      dataWithColor,
+      labels,
+      selectedCategories,
+      theme,
+      showLabel,
+      labelOptions
+    ]
   );
 
   const options = useMemo(
@@ -116,14 +178,41 @@ function PieWidgetUI({
         bottom: theme.spacing(0)
       },
       tooltip: tooltipOptions,
+      legend: legendOptions,
       series: seriesOptions
     }),
-    [theme, tooltipOptions, seriesOptions]
+    [theme, tooltipOptions, seriesOptions, legendOptions]
+  );
+
+  const clickEvent = useCallback(
+    (params) => {
+      if (onSelectedCategoriesChange) {
+        const newSelectedCategories = [...selectedCategories];
+        const { name } = data[params.dataIndex];
+
+        const selectedCategoryIdx = newSelectedCategories.indexOf(name);
+        if (selectedCategoryIdx === -1) {
+          newSelectedCategories.push(name);
+        } else {
+          newSelectedCategories.splice(selectedCategoryIdx, 1);
+        }
+
+        onSelectedCategoriesChange(newSelectedCategories);
+      }
+    },
+    [data, onSelectedCategoriesChange, selectedCategories]
   );
 
   const onEvents = {
-    mouseover: () => setShowLabel(false),
-    mouseout: () => setShowLabel(true)
+    ...(filterable && { click: clickEvent }),
+    mouseover: () => {
+      setShowLabel(false);
+      setShowTooltip(true);
+    },
+    mouseout: () => {
+      setShowLabel(true);
+      setShowTooltip(false);
+    }
   };
 
   return (
@@ -139,8 +228,8 @@ function PieWidgetUI({
 PieWidgetUI.defaultProps = {
   name: null,
   formatter: (v) => v,
-  tooltipFormatter: defaultTooltipFormatter,
-  colors: null,
+  tooltipFormatter,
+  colors: [],
   labels: {},
   height: '260px',
   animation: true,
@@ -168,3 +257,33 @@ PieWidgetUI.propTypes = {
 };
 
 export default PieWidgetUI;
+
+// Aux
+function tooltipFormatter(params) {
+  const value = processFormatterRes(params.formatter(params.value));
+
+  const markerColor = params.data.color || params.textStyle.color;
+  const markerStyle = `display:inline-block;margin-right:4px;border-radius:4px;width:8px;height:8px;background-color:${markerColor}`;
+
+  return `
+    <p style="font-size:12px;font-weight:600;line-height:1.33;margin:4px 0 4px 0;">${params.name}</p>
+    <p style="font-size: 12px;font-weight:normal;line-height:1.33;margin:0 0 4px 0;"><span style="${markerStyle}"></span> ${value} (${params.percent}%)</p>
+  `.trim();
+}
+
+function processDataItem(colorByCategory, colors, theme) {
+  return (item) => {
+    const { name } = item;
+    const colorUsed = colorByCategory.current[name];
+    if (colorUsed) {
+      item.color = colorUsed;
+    } else {
+      const paletteToUse = colors.length ? colors : theme.palette.qualitative.bold;
+      const colorToUse =
+        paletteToUse[Object.keys(colorByCategory.current).length] || '#fff';
+      colorByCategory.current[name] = colorToUse;
+      item.color = colorToUse;
+    }
+    return item;
+  };
+}
